define STACK_CHUNK_SIZE    8192
from the doc:

The stackchunksize parameter does not control the JavaScript stack size. (The JSAPI does not provide a way to adjust the stack depth limit.) Passing a large number for stackchunksize is a mistake. In a DEBUG build, large chunk sizes can degrade performance dramatically. The usual value of 8192 is recommended.

git-svn-id: svn://dyne.org/rastasoft/freej/branches/Goil/freej_91@1053 383723c8-4afa-0310-b8a8-b1afb83214fc