Preliminary 'draft' factory implementation. (WIP)
This first approach involves extending the context to act as a factory
for a set of core classes (at this stage only for Layers but also
Controllers Parameters Encoders and possibly other core-classes could be
handled through the factory).

To decide if we want to keep the factory code bound to the context or if
we prefer the context to use a dedicated factory classes (where the
factory-specific logic will reside).
The second option could be an interesting design pattern ... it would allow
us to subclass the base Factory class to implement the various factories
for the different class hierarchies, instead of forcing us to mantain
different maps and defining different instantiator within the Context class
implementation.
In the meanwhile the first allows me to start testing the factory-based
approach to generalize object creation from javascripts (allowing us, for
example to create an instance of the proper subclas when a statement like
"lay = new VideoLayer()" is executed in the javasacript context)