each script now has its own JSRuntime
various cleanings and fixes while introducing the new
JSExecutionContext internal class (used by the JsParser to
rapresent the execution context of a script).

various fixes and cleanings here and there but more are still needed ...
and at some stage I'd like
to get rid of most of the macros in our internal javascript
abstraction in favour of a more solid api... but first we needed
the internal javascript engine to work properly and allow loading
multiple scripts without having their global context to collide.
We are there now ... but most of the old code needs to be refactored
to take advantage of the new approach ( many js objects still make use
of the global_context/global_object hold by the JsParser)