Now can dump audio and video by the main menu application
In the new AudioCollector constructor, now passes the JackClient in
parameter to be stored.
In JackClient::Process now test if an encoder is connected
before filling the encoder ring buffer.
In OggTheoraEncoder::init now tels Jack there is an encoder.*
Now dumps, but with a big delay between audio and video.