Now can dump in a file with control on audio from the graphic interface
- in QqWidget now creates as many Jack Output ports as there is in
the VideoLayer when QJackClient is launched and none already exists.

- QEncoder no longer herits from OggTheoraEncoder. It can move
in the future. Now deletes on close. Better communication with
QJackClient.

- in QJackClient, now can add Jack Outputs. Now creates the
AudioCollector needed for encoding. Better communication with
QEncoder.

- in JackClient::Process, controls if an encoder running before
filling the encoders ring buffers.

- in JSyncThread::_run, commented out me->_running = true, because
this variable can be accessed from the outside of the thread and
cause problems if stopping a JSyncThread just after created it.
Caused qfreej to hang up when opening creating QEncoder.

- in oggmux_close added some verifications on close to avoid SEGFAULT

- in VideoEncoder destructor, also added verifications.