Now supports audio quality changes when streaming.
- In oggmux_init, sets back the average vorbis bit rate value
in the oggmux_info struct to be able to use it.
Now gets vorbis quality from oggmux_info. Replaced 48000 by oggmux_info
sample_rate.

- In struct oggmux_info, added aveVorBitRate to log the average audio
bit rate value, which depends on vorbis_quality.

- In OggTheoraEncoder::init, now calculates vorbis_quality so that it
is from -0.1 and 1.0 as in vorbis specification when input value is
in between 0 to 100. Added a method to get aveVorBitRate value.

- In QJackClient::init, suppressed the Jack m_SampleRate value
since it is already done by the AudioCollector constructor

- In QEncoder, now gets the bit rate from the OggTheoraEncoder after
launched.